Move communities and community reports
